+++ src/main/java/com/weave/weaveserver/repository/TeamRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface TeamRepository  extends JpaRepository<Team,Integer> {
+public interface TeamRepository  extends JpaRepository<Team, Long> {
 
 }
